Add files via upload
Here's the new robot code with an encoder to measure wheel rotations for autonomous delivery. The next change that needs to be made to it is a system that resets the number of rotations and the autonomous delivery loop when the driver switches back to manual mode. I will be working on the solution to this over the next week, but if anyone has some code they think will work, feel free to upload it.